requirements-ci.txt
fastapi==0.136.0
uvicorn==0.44.0
python-dotenv==1.2.2
requests==2.33.1
groq==0.37.1
langchain==1.2.16
langchain-groq==1.1.2
langchain-core==1.3.2
langchain-community==0.4.1
langgraph==0.4.3
pydantic==2.13.2
